Filled in DBFacade to be easily usable by outside interference
It has a proper constructor and functions. Cannot prove that they work though.
