+++ src/Chirp.CLI/Chirp.CLI.csproj
@@ -1,16 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <PublishSingleFile>true</PublishSingleFile>
-  </PropertyGroup>
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <PublishSingleFile>true</PublishSingleFile>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="docopt.net" Version="0.8.1" />
-    <PackageReference Include="CsvHelper" Version="33.0.1" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="docopt.net" Version="0.8.1"/>
+        <PackageReference Include="CsvHelper" Version="33.0.1"/>
+    </ItemGroup>
 
 </Project>
