+++ src/Chirp.Razor/CheepService.cs
@@ -11,34 +11,20 @@ public interface ICheepService
 
 public class CheepService : ICheepService
 {
-   
-
+    private readonly DBFacade facade;
+    
+    public CheepService(DBFacade facade){
+        this.facade = facade
+    }
 	
     public List<CheepViewModel> GetCheeps()
     {
-        List<CheepViewModel> cheeps = new List<CheepViewModel>();
-        var sqlDBFilePath = "chirp.db";
-        var sqlQuery = @"SELECT u.username, m.text, m.pub_date FROM message m JOIN user u ON m.author_id = u.user_id ORDER by m.pub_date desc";
-        using (var connection = new SqliteConnection($"Data Source={sqlDBFilePath}"))
-        {
-            connection.Open();
-
-            var command = connection.CreateCommand();
-            command.CommandText = sqlQuery;
-
-            using var reader = command.ExecuteReader();
-            while (reader.Read())
-            {
-                var dataRecord = (IDataRecord)reader;
-                cheeps.Add(
-                    new CheepViewModel(
-                        dataRecord[0].ToString(),
-                        dataRecord[1].ToString(), 
-                        UnixTimeStampToDateTimeString(Double.Parse(dataRecord[2].ToString()))));
-                       
-            }
-        }
-        return cheeps;
+        string query = 
+            @"SELECT u.username, m.text, m.pub_date
+            FROM message m
+            JOIN user u ON m.author_id = u.user_id
+            ORDER by m.pub_date desc";
+        return facade.query(query);
     }
 
     public List<CheepViewModel> GetCheepsFromAuthor(string author)
